Move some updates sagas to the main process
